requirements.txt
amqp==2.5.0
astroid==2.2.5
backcall==0.1.0
billiard==3.6.0.0
boto3==1.12.29
botocore==1.15.29
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.2
django-bootstrap4==0.0.8
django-extensions==2.2.3
django-filter==2.2.0
django-storages==1.7.1
django-tables2==2.0.6
docutils==0.14
gunicorn==19.9.0
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.20
jedi==0.13.3
jmespath==0.9.4
kombu==4.6.3
lazy-object-proxy==1.4.1
lxml==4.3.4
mccabe==0.6.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
requests==2.22.0
s3transfer==0.3.3
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
wrapt==1.11.1
